+++ strategies/beaconblockproposal/best/parameters.go
@@ -17,6 +17,7 @@ package best
 
 import (
 	"context"
+	"runtime"
 	"time"
 
 	eth2client "github.com/attestantio/go-eth2-client"
@@ -83,9 +84,10 @@ func WithTimeout(timeout time.Duration) Parameter {
 // parseAndCheckParameters parses and checks parameters to ensure that mandatory parameters are present and correct.
 func parseAndCheckParameters(params ...Parameter) (*parameters, error) {
 	parameters := parameters{
-		logLevel:      zerolog.GlobalLevel(),
-		timeout:       2 * time.Second,
-		clientMonitor: nullmetrics.New(context.Background()),
+		logLevel:           zerolog.GlobalLevel(),
+		timeout:            2 * time.Second,
+		clientMonitor:      nullmetrics.New(context.Background()),
+		processConcurrency: int64(runtime.GOMAXPROCS(-1)),
 	}
 	for _, p := range params {
 		if params != nil {