+++ main.go
@@ -68,7 +68,7 @@ import (
 )
 
 // ReleaseVersion is the release version for the code.
-var ReleaseVersion = "0.6.2"
+var ReleaseVersion = "0.9.0"
 
 func main() {
 	os.Exit(main2())